pyproject.toml
[project]
name = "daisy"
version = "0.2.1"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dashscope>=1.25.17",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pyyaml>=6.0.3",
    "sentence-transformers>=5.4.1",
    "typer>=0.25.0",
    "zvec>=0.3.1",
]

[project.scripts]
daisy = "daisy.cli:app"

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[dependency-groups]
dev = [
    "ruff>=0.15.12",
]
